+++ Runtime/Attributes/Drawers/Grouping/EndGroupAttribute.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using UnityEngine;
 
 namespace RealityProgrammer.OverseerInspector.Runtime.Drawers.Group {
-    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Method, AllowMultiple = true)]
+    [AttributeUsage(AttributeTargets.Field | AttributeTargets.Method | AttributeTargets.Property, AllowMultiple = true)]
     public sealed class EndGroupAttribute : BaseOverseerAttribute {
 
     }